LayoutTests/http/tests/inspector/network/beacon-type-expected.txt

Tests for Resource.Type.Beacon.


== Running test suite: Resource.Type.Beacon
-- Running test case: Resource.Type.Beacon.204
PASS: Resource should be Beacon type.
PASS: Resource should be a POST request.
PASS: Resource should have a 204 response.
PASS: Resource should not have request data.

-- Running test case: Resource.Type.Beacon.404
PASS: Resource should be Beacon type.
PASS: Resource should be a POST request.
PASS: Resource should have a 404 response.
PASS: Resource should not have request data.

-- Running test case: Resource.Type.Beacon.WithData
PASS: Resource should be Beacon type.
PASS: Resource should be a POST request.
PASS: Resource should have a 200 response.
PASS: Resource should have 'SENDING-DATA' request data.
